requirements.txt
discord.py==2.3.2
pytz
Pillow
supabase==2.19.0
python-dotenv==1.0.0
redis>=5.0.0
hiredis>=2.0.0
